+++ src/scvi/data/_manager.py
@@ -216,7 +216,10 @@ class AnnDataManager:
         # If empty, we skip registering the field.
         if not field.is_empty:
             # Transfer case: Source registry is used for validation and/or setup.
-            if source_registry is not None:
+            if (
+                source_registry is not None
+                and field.registry_key in source_registry[_constants._FIELD_REGISTRIES_KEY]
+            ):
                 field_registry[_constants._STATE_REGISTRY_KEY] = field.transfer_field(
                     source_registry[_constants._FIELD_REGISTRIES_KEY][field.registry_key][
                         _constants._STATE_REGISTRY_KEY